jwt claims

+++ resources/item.py
@@ -1,6 +1,6 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -35,6 +35,10 @@ class Items(MethodView):
   
   @jwt_required()
   def delete(self, item_id):
+    jwt = get_jwt()
+    if not jwt.get("is_admin"):
+       abort(401, message="Admin privilage required.")
+
     item = ItemModel.query.get_or_404(item_id)
     db.session.delete(item)
     db.session.commit()
